+++ Sprintfinity3902/Collision/CollisionDetector.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Media;
 using Sprintfinity3902.Entities;
 using Sprintfinity3902.Entities.Enemies_NPCs;
 using Sprintfinity3902.Entities.Items;
@@ -174,6 +175,7 @@ namespace Sprintfinity3902.Collision
             foreach (int enemyID in deletionList)
             {
                 enemies.Remove(enemyID);
+                Sound.SoundLoader.Instance.GetSound(Sound.SoundLoader.Sounds.Boss_Defeated).Play(Global.Var.VOLUME, 0.0f, 0.0f);
             }
         }
