+++ Sprintfinity3902/Content/Content.mgcb
@@ -56,11 +56,29 @@
 /processorParam:Quality=Best
 /build:Sounds/Boss_Zapped.wav
 
+#begin Sounds/Death_Mountain.wav
+/importer:WavImporter
+/processor:SoundEffectProcessor
+/processorParam:Quality=Best
+/build:Sounds/Death_Mountain.wav
+
 #begin Sounds/Door_Opened.wav
 /importer:WavImporter
 /processor:SoundEffectProcessor
 /processorParam:Quality=Best
 /build:Sounds/Door_Opened.wav
+
+#begin Sounds/Dungeon.wav
+/importer:WavImporter
+/processor:SoundEffectProcessor
+/processorParam:Quality=Best
+/build:Sounds/Dungeon.wav
+
+#begin Sounds/Ending.wav
+/importer:WavImporter
+/processor:SoundEffectProcessor
+/processorParam:Quality=Best
+/build:Sounds/Ending.wav
 
 #begin Sounds/Enemy_Shrinking.wav
 /importer:WavImporter
@@ -80,11 +98,29 @@
 /processorParam:Quality=Best
 /build:Sounds/Flames_Shot.wav
 
+#begin Sounds/Game_Over.wav
+/importer:WavImporter
+/processor:SoundEffectProcessor
+/processorParam:Quality=Best
+/build:Sounds/Game_Over.wav
+
 #begin Sounds/Health_Heart.wav
 /importer:WavImporter
 /processor:SoundEffectProcessor
 /processorParam:Quality=Best
 /build:Sounds/Health_Heart.wav
+
+#begin Sounds/Intro.wav
+/importer:WavImporter
+/processor:SoundEffectProcessor
+/processorParam:Quality=Best
+/build:Sounds/Intro.wav
+
+#begin Sounds/Item.wav
+/importer:WavImporter
+/processor:SoundEffectProcessor
+/processorParam:Quality=Best
+/build:Sounds/Item.wav
 
 #begin Sounds/Item_Obtained.wav
 /importer:WavImporter
@@ -128,6 +164,12 @@
 /processorParam:Quality=Best
 /build:Sounds/Master_Sword_Sword_Zap.wav
 
+#begin Sounds/Overworld.wav
+/importer:WavImporter
+/processor:SoundEffectProcessor
+/processorParam:Quality=Best
+/build:Sounds/Overworld.wav
+
 #begin Sounds/Player_Hurt.wav
 /importer:WavImporter
 /processor:SoundEffectProcessor
@@ -145,6 +187,12 @@
 /processor:SoundEffectProcessor
 /processorParam:Quality=Best
 /build:Sounds/Rupees_Decreasing.wav
+
+#begin Sounds/Secret.wav
+/importer:WavImporter
+/processor:SoundEffectProcessor
+/processorParam:Quality=Best
+/build:Sounds/Secret.wav
 
 #begin Sounds/Shield_Deflecting.wav
 /importer:WavImporter
@@ -170,6 +218,12 @@
 /processorParam:Quality=Best
 /build:Sounds/The_Triforce_Reveals_Ganon.wav
 
+#begin Sounds/Triforce.wav
+/importer:WavImporter
+/processor:SoundEffectProcessor
+/processorParam:Quality=Best
+/build:Sounds/Triforce.wav
+
 #begin Sounds/Triforce_Piece_Obtained.wav
 /importer:WavImporter
 /processor:SoundEffectProcessor
@@ -193,6 +247,12 @@
 /processor:SoundEffectProcessor
 /processorParam:Quality=Best
 /build:Sounds/Walking_on_Stairs.wav
+
+#begin Sounds/Whistle.wav
+/importer:WavImporter
+/processor:SoundEffectProcessor
+/processorParam:Quality=Best
+/build:Sounds/Whistle.wav
 
 #begin Zelda_Bosses_Transparent.png
 /importer:TextureImporter
