+++ Sprintfinity3902/Global/Var.cs
@@ -3,5 +3,6 @@
     public static class Var
     {
         public static int SCALE = 4;
+        public static float VOLUME = 0.01f;
     }
 }
